core/requirements.txt
numpy
psycopg2-binary
netCDF4
python-dotenv
fastapi
fastmcp